// apps/lottery-domain/src/strategy/model/entity/StrategyFlowRecordEntity.ts
export class StrategyFlowRecordEntity {
    userId: string;
    strategyId: number;
    orderId: string;
    awardId: number;
    nodeDesc: string;
    processType: string;
    chainProcessResult: string;
    treeProcessResult: string;
    ruleLimitValue: string;
    currentNode: string;
    nextNode: string;
    head: number;
    treeId: string;
}